added register request service in UserController.java and added the dto package along with the User.java object
