--- src/main/java/com/redcom/redcom/User.java
@@ -1,45 +1,0 @@
-package com.redcom.redcom;
-
-public class User {
-
-    private int id;
-
-    private String username;
-
-    private String password;
-
-    private String email;
-
-    public User(int id, String username, String password, String email)
-    {
-        this.id = id;
-        this.username = username;
-        this.password = password;
-        this.email = email;
-    }
-
-    public int getId()
-    {
-        return id;
-    }
-
-    public int setId(int id)
-    {
-        return id;
-    }
-
-    public String getUsername()
-    {
-        return username;
-    }
-
-    public String getPassword()
-    {
-        return password;
-    }
-
-    public String getEmail()
-    {
-        return email;
-    }
-}
